refactor(commands): extract server tests to sibling file
